requirements.txt
Flask==2.3.3
google-generativeai==0.1.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==23.0.0